Refactor towards pre-algebra-page
